QA: Split off each class into its own file.
