+++ class-debug-bar-constants.php
@@ -1,12 +1,10 @@
 <?php
 /**
- * Debug Bar Constants - Debug Bar Panels.
+ * Debug Bar Constants - Base class for a Debug Bar Constants Debug Bar Panel.
  *
  * @package     WordPress\Plugins\Debug Bar Constants
  * @author      Juliette Reinders Folmer <wpplugins_nospam@adviesenzo.nl>
  * @link        https://github.com/jrfnl/Debug-Bar-Constants
- * @since       1.0
- * @version     1.6.1
  *
  * @copyright   2013-2016 Juliette Reinders Folmer
  * @license     http://creativecommons.org/licenses/GPL/2.0/ GNU General Public License, version 2 or higher
@@ -20,9 +18,6 @@ if ( ! function_exists( 'add_action' ) ) {
 }
 
 
-/**
- * The classes in this file extend the functionality provided by the parent plugin "Debug Bar".
- */
 if ( ! class_exists( 'Debug_Bar_Constants' ) && class_exists( 'Debug_Bar_Panel' ) ) {
 
 	/**
@@ -136,232 +131,5 @@ if ( ! class_exists( 'Debug_Bar_Constants' ) && class_exists( 'Debug_Bar_Panel' 
 			}
 		}
 	} // End of class Debug_Bar_Constants.
-
-} // End of if class_exists wrapper.
-
-
-if ( ! class_exists( 'Debug_Bar_WP_Constants' ) && class_exists( 'Debug_Bar_Constants' ) ) {
-
-	/**
-	 * Debug Bar WP Constants.
-	 */
-	class Debug_Bar_WP_Constants extends Debug_Bar_Constants {
-
-
-		/**
-		 * Constructor.
-		 */
-		public function init() {
-			parent::init();
-			$this->title( __( 'WP Constants', 'debug-bar-constants' ) );
-		}
-
-
-		/**
-		 * Limit visibility of the output to super admins on multi-site and
-		 * admins on non multi-site installations.
-		 */
-		public function prerender() {
-			$this->set_visible( is_super_admin() );
-		}
-
-
-		/**
-		 * Render the output.
-		 */
-		public function render() {
-			$constants = get_defined_constants( true );
-			if ( isset( $constants['user'] ) && ( ! empty( $constants['user'] ) && is_array( $constants['user'] ) ) ) {
-				echo '
-		<h2><span>', esc_html__( 'Constants within WP:', 'debug-bar-constants' ), '</span>', absint( count( $constants['user'] ) ), '</h2>';
-				$this->dbc_render_table( $constants['user'] );
-
-			} else {
-				// Should never happen.
-				echo '<p>', esc_html__( 'No constants found... this is really weird...', 'debug-bar-constants' ), '</p>';
-			}
-		}
-	} // End of class Debug_Bar_WP_Constants.
-
-} // End of if class_exists wrapper.
-
-
-if ( ! class_exists( 'Debug_Bar_WP_Class_Constants' ) && class_exists( 'Debug_Bar_Constants' ) ) {
-
-	/**
-	 * Debug Bar WP Class Constants.
-	 */
-	class Debug_Bar_WP_Class_Constants extends Debug_Bar_Constants {
-
-
-		/**
-		 * Constructor.
-		 */
-		public function init() {
-			parent::init();
-			$this->title( __( 'WP Class Constants', 'debug-bar-constants' ) );
-		}
-
-		/**
-		 * Render the output.
-		 */
-		public function render() {
-
-			$classes = get_declared_classes();
-			if ( class_exists( 'Debug_Bar_List_PHP_Classes' ) && property_exists( 'Debug_Bar_List_PHP_Classes', 'PHP_classes' ) ) {
-				$classes = array_udiff( $classes, Debug_Bar_List_PHP_Classes::$PHP_classes, 'strcasecmp' );
-			}
-
-			$constants = array();
-
-			if ( ! empty( $classes ) && is_array( $classes ) ) {
-				// Get the constants info first.
-				foreach ( $classes as $class ) {
-					$reflector       = new ReflectionClass( $class );
-					$class_constants = $reflector->getConstants();
-
-					if ( ! empty( $class_constants ) && is_array( $class_constants ) ) {
-						$constants[ $class ] = $class_constants;
-					}
-					unset( $class_constants, $reflector );
-				}
-				unset( $class );
-
-				// Generate the output.
-				if ( ! empty( $constants ) && is_array( $constants ) ) {
-					uksort( $constants, 'strnatcasecmp' );
-
-					foreach ( $constants as $class => $set ) {
-						$count = count( $set );
-						echo '
-			<h2><a href="#dbcwpc-', esc_attr( $class ), '"><span>', esc_html( $class ), ':</span>', absint( $count ), '</a></h2>';
-						unset( $count );
-					}
-					unset( $class, $set );
-
-					echo '<p class="dbcwpc-info">', wp_kses_post( __( '<strong>Please note</strong>: these may be both native WordPress classes as well as classes which may be declared by plugins or themes.<br />You can use these constants in your code using <code>class_name::constant_name</code>.', 'debug-bar-constants' ) ), ' ',
-					/* TRANSLATORS: %s = the "href" element for the link. */
-					wp_kses_post( sprintf( __( 'See the <a %s>FAQ</a> for more information.', 'debug-bar-constants' ), 'href="https://wordpress.org/plugins/debug-bar-constants/faq/" target="_blank"' ) ), '.</p>';
-
-					foreach ( $constants as $class => $set ) {
-						echo '
-			<h3 id="dbcwpc-', esc_attr( $class ), '"><em>', esc_html( ucfirst( $class ) ), '</em> ', esc_html__( 'Constants:', 'debug-bar-constants' ), '</h3>';
-						$this->dbc_render_table( $set );
-					}
-					unset( $class, $set );
-				}
-			} else {
-				// Should never happen.
-				echo '<p>', esc_html__( 'No classes nor class constants found... this is kinda strange...', 'debug-bar-constants' ), '</p>';
-			}
-		}
-	} // End of class Debug_Bar_WP_Class_Constants.
-
-} // End of if class_exists wrapper.
-
-
-if ( ! class_exists( 'Debug_Bar_PHP_Constants' ) && class_exists( 'Debug_Bar_Constants' ) ) {
-
-	/**
-	 * Debug Bar PHP Constants.
-	 */
-	class Debug_Bar_PHP_Constants extends Debug_Bar_Constants {
-
-
-		/**
-		 * Constructor.
-		 */
-		public function init() {
-			parent::init();
-			$this->title( __( 'PHP Constants', 'debug-bar-constants' ) );
-		}
-
-		/**
-		 * Render the output.
-		 */
-		public function render() {
-
-			$constants = get_defined_constants( true );
-			unset( $constants['user'] );
-
-			if ( ! empty( $constants ) && is_array( $constants ) ) {
-				uksort( $constants, 'strnatcasecmp' );
-
-				foreach ( $constants as $category => $set ) {
-					echo '
-		<h2><a href="#dbcphp-', esc_attr( $category ), '"><span>', esc_html( $category ), ':</span>', absint( count( $set ) ), '</a></h2>';
-				}
-				unset( $category, $set );
-
-				foreach ( $constants as $category => $set ) {
-					if ( ! empty( $set ) && is_array( $set ) ) {
-
-						// Set url to correct page in the PHP manual for more info.
-						$url = $this->get_php_manual_url( $category );
-
-						/* TRANSLATORS: %s = the name of a PHP extension. */
-						$title_attr = sprintf( __( 'Visit the PHP manual page about the %s constants.', 'debug-bar-constants' ), $category );
-
-						echo '
-		<h3 id="dbcphp-', esc_attr( $category ), '"><em><a href="', esc_url( $url ), '" target="_blank" title="', esc_attr( $title_attr ), '">', esc_html( ucfirst( $category ) ), '</a></em> ', esc_html__( 'Constants:', 'debug-bar-constants' ), '</h3>';
-						$this->dbc_render_table( $set );
-					}
-				}
-				unset( $category, $set, $title_attr );
-
-			} else {
-				// Should never happen.
-				echo '<p>', esc_html__( 'No PHP constants found... this is really weird...', 'debug-bar-constants' ), '</p>';
-			}
-		}
-
-
-		/**
-		 * Retrieve the PHP manual URL for the constants page of a specific PHP extension.
-		 *
-		 * Works round some of the peculiarities of the PHP.net URL scheme.
-		 *
-		 * @param string $category The PHP Extension for which to retrieve the URL.
-		 *
-		 * @return string URL
-		 */
-		protected function get_php_manual_url( $category ) {
-			switch ( $category ) {
-				case 'Core':
-					$url = 'http://php.net/reserved.constants';
-					break;
-
-				case 'date':
-					$url = 'http://php.net/datetime.constants';
-					break;
-
-				case 'gd':
-					$url = 'http://php.net/image.constants';
-					break;
-
-				case 'odbc':
-					$url = 'http://php.net/uodbc.constants';
-					break;
-
-				case 'standard':
-					$url = ''; // Definitions are all over, part of core.
-					break;
-
-				case 'tokenizer':
-					$url = 'http://php.net/tokens';
-					break;
-
-				case 'xdebug':
-					$url = 'http://xdebug.com/docs/';
-					break;
-
-				default:
-					$url = 'http://php.net/' . rawurlencode( $category ) . '.constants';
-					break;
-			}
-
-			return $url;
-		}
-	} // End of class Debug_Bar_PHP_Constants.
 
 } // End of if class_exists wrapper.
